requirements.txt
ibm-watson
pandas
pdfplumber
docx2txt
streamlit
python-dotenv
Flask
